+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## CHANGE LOG.
 
+### October 27, 2023
+
+* [CleverTap Push Templates SDK v1.2.0](docs/CTPUSHTEMPLATESCHANGELOG.md).
+
+### October 12, 2023
+
+* [CleverTap Android SDK v5.2.1](docs/CTCORECHANGELOG.md)
+* [CleverTap Xiaomi Push SDK v1.5.4](docs/CTXIAOMIPUSHCHANGELOG.md)
+
 ### August 10, 2023
 
 * [CleverTap Android SDK v5.2.0](docs/CTCORECHANGELOG.md)